+++ gateway/Routers/shopRouters.py
@@ -14,30 +14,34 @@ def create_vent():
     json = {}
 
     # comprobar si existe el producto
-    headers = {"Content-Type": "application/json; charset=utf-8"}
+    headers = {
+        'Authorization':request.headers.get('Authorization'),
+        "Content-Type": "application/json; charset=utf-8"}
     data = request.get_json()
     url = "http://" + dataConfig["url-backend"] + ":" + str(dataConfig["port"])
     dontSave = []
     for prd in data["prd"]:
         if data["empresa"] == prd["idempresa"]:
-            urlp = url + "/articulo_id/" + str(prd["idproducto"])
+            urlp = url + "/articulo_id/" + str(prd["idproducto"])+"/fact/"+str(prd["idempresa"])
             json = {"idEmpresa": prd["idempresa"]}
             response = requests.get(urlp, headers=headers, json=json)
-            producto = response.json()["Producto"]
-            if producto is None:
-                dontSave.append(prd)
-            else:
-                if int(producto["prdinfo"]["stock"]) <= prd["cantidad"]:
+            if response.json()["successful"]:
+                producto = response.json()["Producto"]
+                if producto is None:
                     dontSave.append(prd)
-                prd["precioUn"] = producto["prdinfo"]["costoVenta"]
+                else:
+                    if int(producto["prdinfo"]["stock"]) <= prd["cantidad"]:
+                        dontSave.append(prd)
+                    prd["precioUn"] = producto["prdinfo"]["costoVenta"]
+            else:
+                dontSave.append(prd)
         else:
             dontSave.append(prd)
-
     for dat in dontSave:
         data["prd"].remove(dat)
 
     if data["prd"]:
-        url = dataConfig["url-backend-shopp"] + "/venta"
+        url = dataConfig["url-backend-shopp"] + "/venta_create/"+data["empresa"]
         response = requests.post(url, headers=headers, json=data)
         datos=response.json()
         successful=True
@@ -56,20 +60,26 @@ def create_vent():
 #        obtienen uno solo
 
 # obtiene un producto vendido por id
-@ventas.route("/vent_prod_id/<id>", methods=['GET'])
-def get_vent_by_id(id):
+@ventas.route("/vent_prod_id/<id>/empr/<idEmpr>", methods=['GET'])
+def get_vent_by_id(id, idEmpr):
     headers = {"Content-Type": "application/json; charset=utf-8"}
-    url = dataConfig["url-backend-shopp"] + "/venta_product_id/" + id
-    response = requests.get(url, headers=headers)
+    if comprobar_empresa(idEmpr):
+        url = dataConfig["url-backend-shopp"] + "/venta_product_id/" + id +"/empr/"+idEmpr
+        response = requests.get(url, headers=headers)
+    else:
+        return "False"
     return response.json()
 
 
 # obtiene una venta en facturacion de la empresa por su id
-@ventas.route("/vent_fact_id/<id>", methods=['GET'])
-def get_fact_pro_by_id_emp(id):
+@ventas.route("/vent_fact_id/<id>/empr/<empresa>", methods=['GET'])
+def get_fact_pro_by_id_emp(id, empresa):
     headers = {"Content-Type": "application/json; charset=utf-8"}
-    url = dataConfig["url-backend-shopp"] + "/get_only_fact_id/" + id
-    response = requests.get(url, headers=headers)
+    if comprobar_empresa(empresa):
+        url = dataConfig["url-backend-shopp"] + "/get_only_fact_id/" + id + "/empr/"+empresa
+        response = requests.get(url, headers=headers)
+    else:
+        return "False"
     return response.json()
 
 
@@ -79,8 +89,11 @@ def get_fact_pro_by_id_emp(id):
 @ventas.route("/vent_prod_id_empr/<id>", methods=['GET'])
 def get_vent_pro_by_id_emp(id):
     headers = {"Content-Type": "application/json; charset=utf-8"}
-    url = dataConfig["url-backend-shopp"] + "/get_prod_id_em/" + id
-    response = requests.get(url, headers=headers)
+    if comprobar_empresa(id):
+        url = dataConfig["url-backend-shopp"] + "/get_prod_id_em/" + id
+        response = requests.get(url, headers=headers)
+    else:
+        return "False"
     return response.json()
 
 
@@ -88,16 +101,22 @@ def get_vent_pro_by_id_emp(id):
 @ventas.route("/vent_id_empresa/<empresa>", methods=['GET'])
 def get_vent_by_id_empr(empresa):
     headers = {"Content-Type": "application/json; charset=utf-8"}
-    url = dataConfig["url-backend-shopp"] + "/venta_facturacion_id/" + empresa
-    response = requests.get(url, headers=headers)
+    if comprobar_empresa(empresa):
+        url = dataConfig["url-backend-shopp"] + "/venta_facturacion_id/" + empresa
+        response = requests.get(url, headers=headers)
+    else:
+        return "False"
     return response.json()
 
 
-@ventas.route("/delete_vent_id_empresa/<empresa>", methods=['DELETE'])
-def delete_vent_by_id_empr(empresa):
+@ventas.route("/delete_vent_id_empresa/<id>/empr/<idEmp>", methods=['DELETE'])
+def delete_vent_by_id_empr(id,idEmp):
     headers = {"Content-Type": "application/json; charset=utf-8"}
-    url = dataConfig["url-backend-shopp"] + "/delete_facturacion_id/" + empresa
-    response = requests.delete(url, headers=headers)
+    if comprobar_empresa(idEmp):
+        url = dataConfig["url-backend-shopp"] + "/delete_facturacion_id/" + id+"/empr/"+idEmp
+        response = requests.delete(url, headers=headers)
+    else:
+        return "False"
     return response.json()
 
 
@@ -108,3 +127,18 @@ def get_all_vent():
     url = dataConfig["url-backend-shopp"] + "/facturas"
     response = requests.get(url, headers=headers)
     return response.json()
+
+
+
+
+
+
+
+def comprobar_empresa(idempresa):
+    headers = {
+        'Authorization': request.headers.get('Authorization'),
+        "Content-Type": "application/json; charset=utf-8"
+    }
+    verificarEmpresa = dataConfig["url-backend-security"] + "/factory/getByid/" + idempresa
+    response = requests.get(verificarEmpresa, headers=headers)
+    return response.json()["seccessful"]
